+++ src/com/xxxx/entity/CompanyForm.java
@@ -15,14 +15,14 @@ public class CompanyForm {
     Double Raise_to_Date;
     Integer Employee_No;
     Double Revenue;
-    Integer IRR;
+    Double IRR;
     // Valuation fields
     Double Post_Valuation;
     String Valuation_Change_Reason;
     Double MSEQ_Investment_Cur_Val;
     Double Own_Percent;
 
-    public CompanyForm(String company_Name, String theme, Integer year_Founded, String runway_Start_Date, String runway_End_Date, Integer runway_Month, Double raise_to_Date, Integer employee_No, Double revenue, Integer IRR, Double post_Valuation, String valuation_Change_Reason, Double MSEQ_Investment_Cur_Val, Double own_Percent) {
+    public CompanyForm(String company_Name, String theme, Integer year_Founded, String runway_Start_Date, String runway_End_Date, Integer runway_Month, Double raise_to_Date, Integer employee_No, Double revenue, Double IRR, Double post_Valuation, String valuation_Change_Reason, Double MSEQ_Investment_Cur_Val, Double own_Percent) {
         Company_Name = company_Name;
         Theme = theme;
         Year_Founded = year_Founded;
